LAB 5

+++ users-service/src/main/java/com/hrworkflow/usersservice/repository/UserRepository.java
@@ -9,11 +9,11 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByEmail(String email);
     List<User> findByRole(Role role);
     List<User> findByPosition(String position);
 
-    boolean existsByIdAndRoleIs(Integer candidateId, Role role);
+    boolean existsByIdAndRoleIs(Long candidateId, Role role);
 }
